Re-organize actions modules
