+++ tests/test.purchase.jsx
@@ -5,9 +5,9 @@ var TestUtils;
 var assign = require('object-assign');
 var rewire = require('rewire');
 
-var actionTypes = require('action-types');
+var actionTypes = require('actions/types');
 var reduxConfig = require('redux-config');
-var purchaseActions = require('purchase-actions');
+var purchaseActions = require('actions/purchase');
 
 var helpers = require('./helpers');
